Move CCBR notation parsing
